+++ views/list_historiales.php
@@ -16,17 +16,24 @@
   <?php include '../views/navbar.php';?>
 <div class="container">
 </div>
+
+<?php  if (( $_SESSION['rol'] == "archivador")||( $_SESSION['rol'] == "admin"))   {  ?>
+
    <center> <th><a  href="../views/add_historial.php"  class="btn btn-primary" role="button" >Agregar</a></th></center>
+<?php } ?>
+
+
 <div class="container">
   <h2>Historiales </h2>
-  <p>historiales del sistemas .</p>
+  <p>En las historias vas poder encontrar los pedidos que realizaste y las entregas con tu usuario .</p>
   <table class="table">
     <thead>
       <tr>
         <th>usuario</th>
         <th>codigo</th>
         <th>estado</th>
-        <th>cod</th>
+        <th>fecha</th>
+      
       </tr>
     </thead>
     <tbody>
@@ -36,10 +43,22 @@
         <td><?php echo  $h["id_usuario"] ?></td>
         <td><?php echo  $h["id_caja"] ?></td>
         <td><?php echo  $h["estado"] ?></td>
-        <td><img alt="Coding Sips" src="../barcode/barcode.php?text=<?php echo  $h["codigo"]?>&print=false" /></td>
+        <td><?php echo  $h["fecha"] ?></td>
+
+
+        
+       
+<?php  if (( $_SESSION['rol'] == "archivador")||( $_SESSION['rol'] == "admin"))   {  ?>
+
+    <td><img alt="Coding Sips" src="../barcode/barcode.php?text=<?php echo  $h["codigo"]?>&print=false" /></td>
         <th><a href="../views/show_historial.php?id_historial=<?php echo  $h["id_historial"] ?>"  class="btn btn-info" role="button" >Informacion</a></th>
         <th><a  href="../views/delete_historial.php?id_historial=<?php echo  $h["id_historial"] ?>"  class="btn btn-danger" role="button" >Eliminar</a></th>
         <th><a  href="../views/edit_historial.php?id_historial=<?php echo  $h["id_historial"] ?>"  class="btn btn-warning" role="button" >Modificar</a></th>
+<?php } ?>
+
+       
+
+
   
   <?php } ?>
       </tr>
